Rounding fixes, and link tax rows to tax rates by ID.
Also closes #1869 as it allows us to use 'codes' for tax rates instead
of their label. E.g. code US-AL-Tax-1. This lets users filter taxes in
reports easier.

+++ woocommerce-core-functions.php
@@ -621,7 +621,7 @@ function woocommerce_trim_zeros( $price ) {
  * @return string
  */
 function woocommerce_format_decimal( $number ) {
-	$number = number_format( (float) $number, get_option( 'woocommerce_price_num_decimals' ), '.', '' );
+	$number = number_format( (float) $number, (int) get_option( 'woocommerce_price_num_decimals' ), '.', '' );
 
 	if ( strstr( $number, '.' ) )
 		$number = rtrim( rtrim( $number, '0' ), '.' );
@@ -638,7 +638,7 @@ function woocommerce_format_decimal( $number ) {
  * @return float
  */
 function woocommerce_format_total( $number ) {
-	return number_format( (float) $number, get_option( 'woocommerce_price_num_decimals' ), '.', '' );
+	return number_format( (float) $number, (int) get_option( 'woocommerce_price_num_decimals' ), '.', '' );
 }
